requirements.txt
modal
requests
pydantic==2.4.2
instructor
pydub
python-telegram-bot
openai